Updates to BTree/Node & TreeObject
Set up loop in insertion method to properly iterate through current node and will: insert (if leaf), call for split (if full), increment frequency (if exact double found), or call for the child (if not leaf, exact value not found, but is in-between two values of current node). Work in progress, forced to stop early for the night (see TODO marker). Do not have the read/write methods implemented, though they have been tested previously.
Corrected order of insertion in BTree - was checking from end to front for larger objects, when it should have been looking for first smaller.
Corrected incrementation of 'n' node object count variable in BTreeNode class - was not being incremented and decremented appropriately when objects added or removed.